rt-tests: Handle -lcpupower flag outside LDFLAGS
The rt-tests Makefile adds -lcpupower to LDFLAGS if libcpupower support
is detected on the system. However, this does not work when LDFLAGS is
overwritten by an external source.

Add separate variable LIBCPUPOWER to contain the flag and pass it to the
cyclictest linking phase. This also avoids unnecessarily passing the
flag to the linking of other rt-tests binaries which do not use
libcpupower.

Signed-off-by: Tomas Glozar <tglozar@redhat.com>
Signed-off-by: John Kacur <jkacur@redhat.com>
(cherry picked from commit 623f6908ef3bc8367661a2e1440dd8c5bee91397)
Signed-off-by: Anders Roxell <anders.roxell@linaro.org>